refactor: extract TestConfiguration to own file (Phase 7)
- Move TestConfiguration record to dedicated file
- Change visibility from private to package-private
- Add comprehensive Javadoc documentation
- ~113 lines removed from ProbabilisticTestExtension (1132 → 1019)
