+++ src/main/java/org/javai/punit/ptest/engine/ProbabilisticTestExtension.java
@@ -979,119 +979,6 @@ public class ProbabilisticTestExtension implements
 	// ========== Inner Classes ==========
 
 	/**
-	 * Holds the resolved test configuration.
-	 */
-	private record TestConfiguration(
-			int samples,
-			double minPassRate,
-			double appliedMultiplier,
-			long timeBudgetMs,
-			int tokenCharge,
-			long tokenBudget,
-			CostBudgetMonitor.TokenMode tokenMode,
-			BudgetExhaustedBehavior onBudgetExhausted,
-			ExceptionHandling onException,
-			int maxExampleFailures,
-			// Statistical context for failure messages (null for legacy/spec-less mode)
-			Double confidence,
-			Double baselineRate,
-			Integer baselineSamples,
-			String specId,
-			// Pacing configuration
-			PacingConfiguration pacing,
-		// Transparent stats configuration
-		TransparentStatsConfig transparentStats,
-		// Provenance metadata
-		ThresholdOrigin thresholdOrigin,
-		String contractRef
-) {
-		boolean hasMultiplier() {
-			return appliedMultiplier != 1.0;
-		}
-
-		boolean hasTimeBudget() {
-			return timeBudgetMs > 0;
-		}
-
-		boolean hasTokenBudget() {
-			return tokenBudget > 0;
-		}
-
-		boolean hasPacing() {
-			return pacing != null && pacing.hasPacing();
-		}
-
-		boolean hasStatisticalContext() {
-			return confidence != null && baselineRate != null && baselineSamples != null && specId != null;
-		}
-
-		boolean hasTransparentStats() {
-			return transparentStats != null && transparentStats.enabled();
-		}
-
-		/**
-		 * Returns true if any provenance information is specified.
-		 */
-		boolean hasProvenance() {
-			return hasThresholdOrigin() || hasContractRef();
-		}
-
-		/**
-		 * Returns true if thresholdOrigin is specified (not UNSPECIFIED).
-		 */
-		boolean hasThresholdOrigin() {
-			return thresholdOrigin != null && thresholdOrigin != ThresholdOrigin.UNSPECIFIED;
-		}
-
-		/**
-		 * Returns true if contractRef is specified (not null or empty).
-		 */
-		boolean hasContractRef() {
-			return contractRef != null && !contractRef.isEmpty();
-		}
-
-		/**
-		 * Creates a copy of this configuration with an updated minPassRate.
-		 *
-		 * <p>Used when the minPassRate is derived from a baseline after lazy selection.
-		 */
-		TestConfiguration withMinPassRate(double newMinPassRate) {
-			return new TestConfiguration(
-					samples, newMinPassRate, appliedMultiplier, timeBudgetMs, tokenCharge, tokenBudget,
-					tokenMode, onBudgetExhausted, onException, maxExampleFailures,
-					confidence, baselineRate, baselineSamples, specId,
-					pacing, transparentStats, thresholdOrigin, contractRef
-			);
-		}
-
-		/**
-		 * Builds the statistical context for failure messages.
-		 */
-		PunitFailureMessages.StatisticalContext buildStatisticalContext(
-				double observedRate, int successes, int samplesExecuted) {
-			if (hasStatisticalContext()) {
-				return new PunitFailureMessages.StatisticalContext(
-						confidence,
-						observedRate,
-						successes,
-						samplesExecuted,
-						minPassRate,
-						baselineRate,
-						baselineSamples,
-						specId
-				);
-			} else {
-				return PunitFailureMessages.StatisticalContext.forLegacyMode(
-						observedRate,
-						successes,
-						samplesExecuted,
-						minPassRate
-				);
-			}
-		}
-	}
-
-	/**
 	 * Invocation context for a single sample execution.
 	 * Provides ParameterResolver for TokenChargeRecorder injection.
 	 */
